BACKLOG-20632: Fix PR feedbacks.
