--- core/src/javascript/components/contentEditor/VanityUrlTableData/index.js
@@ -1,3 +1,0 @@
-import VanityUrlTableData from './VanityUrlTableData';
-
-export default VanityUrlTableData;
