+++ core/src/javascript/components/gqlQueries.js
@@ -2,7 +2,7 @@ import {DefaultVanityUrls, LiveVanityUrls} from './gqlFragments';
 import {PredefinedFragments} from '@jahia/data-helper';
 import gql from 'graphql-tag';
 
-const TableQuery = gql`
+const DashboardTableQuery = gql`
     query NodesQuery($lang: String!, $offset: Int, $limit: Int, $query: String!, $filterText: String, $doFilter: Boolean!, $queryFilter: InputFieldFiltersInput, $languages: [String!]) {
         jcr {
             nodesByQuery(query: $query, limit: $limit, offset: $offset, fieldFilter: $queryFilter) {
@@ -12,6 +12,26 @@ const TableQuery = gql`
                 nodes {
                     ...NodeCacheRequiredFields
                     mixinTypes{
+                        name
+                    }
+                    displayName(language: $lang)
+                    ...DefaultVanityUrls
+                    ...LiveVanityUrls
+                }
+            }
+        }
+    }
+    ${DefaultVanityUrls}
+    ${LiveVanityUrls}
+`;
+
+const ContentEditorTableQuery = gql`
+    query NodesQuery($query: String!, $lang: String!, $filterText: String = "", $doFilter: Boolean!, $languages: [String!], $queryFilter: InputFieldFiltersInput) {
+        jcr {
+            nodesByQuery(query: $query, queryLanguage: XPATH, fieldFilter: $queryFilter) {
+                nodes {
+                    ...NodeCacheRequiredFields
+                   mixinTypes{
                         name
                     }
                     displayName(language: $lang)
@@ -60,4 +80,4 @@ const GetNodeQuery = gql`
     ${PredefinedFragments.nodeCacheRequiredFields.gql}
 `;
 
-export {TableQuery, GetNodeQuery, VanityUrlsByPath, VanityUrlsByPathVariables};
+export {DashboardTableQuery, ContentEditorTableQuery, GetNodeQuery, VanityUrlsByPath, VanityUrlsByPathVariables};
